refactor(contact): server page with client form island
